Fail doc/api generation when a private super class is exposed.
Added a new error type PRIVATE_SUPERCLASS, which
fails the doc/api generation when a public class's super class is private
and not marked as "@hide".
Note the difference between PRIVATE_SUPERCLASS and the existing
HIDDEN_SUPERCLASS. See their usage in Stubs.java.
HIDDEN_SUPERCLASS: a super class is marked as "@hide". Currently doclava
strips such a class from the doc/api file and gives a warning by default;
PRIVATE_SUPERCLASS: a super class is private but not marked as "@hide".
Previously doclava just outputs the private class to doc/api file. But the
generated api file will fail apicheck. With this change, we fail the doc/api
generation in that case.

Bug:19994527
Change-Id: I7971e40bf7280187253d003cc90ab80d13c986c4