Add support to doclava for default / static interface methods
This change removes code that forces all interface methods to be
abstract. It also passes the new "is default" information to all
the places that require it.

Includes changes to the doclava template. Ordering of
method declaration modifiers changed slightly to reflect the ordering
given in:

http://docs.oracle.com/javase/specs/jls/se8/html/jls-8.html#jls-8.4.3
http://docs.oracle.com/javase/specs/jls/se8/html/jls-9.html#jls-9.4

This change *requires* the JDK 8 version of javadoc.

Bug: 26241119
Change-Id: Iccd3006ab52585cc7529e4924bbeea9aef36f046