Add parameter type to method documentation. This rewrites the paramTags() method to use just one loop based on the ParameterInfo for each method parameter. That is, for each parameter identified from the method signature, collect the name and type... then compare that name to the names in the @param tags. If one matches, then collect that @param documentation.
This is the inverse of the way it used to be, which collected the
parameter documentation based soley on what was given in the @param
tags. So any parameters without the @param tag would simply not be
documented in the output... Now, all parameters get listed with their
name and type. The comments from @param are considered a bonus.

Unfortunately, this does not create links to the parameter type.
That will require a bit more work to change the way the data gets
written to the HDF data. So I decided to do this part as an incremental
improvement. Another revision to allow linking to data types will come later.

Change-Id: I32b26ba53c0637d4f5c04eee67bf515e0004a529