Fix false positive for simultaneously hidden and non hidden hierarchies.
Historically doclava has assumed that if the method that was
immediately overriden was hidden (or its containing class was), then
its whole hierarchy was hidden. This isn't true in the case where we
have a non-hidden class with methods that are overriden by hidden
subclasses and then overriden again in a non-hidden subclass of
theirs.

Bug: 28793355
Change-Id: Ib714b7547bd223223aa5499e4752d8c98205334a